refactor(content): extract GroupHeader and section-grouping helper into its own file
